refactor: decouple ecosystem from existing code base (#764)
* feat: decouple ecosystem codebase


* feat: org dashboard changes


* fix: added conditions for empty list components and create endorsement buttons


* [Deno Deploy] Update .github/workflows/deploy.yml

* feat:added env varibales in envConfig


* feat: added ecosystem intercepter for ecosystem


* refactor: changes in axios methods


* fix: redirection logic for create schema page


* fix: delete organization issues


* refactor: removed unused and commented code


---------
