+++ .github/workflows/deploy.yml
@@ -1,9 +1,9 @@
 name: Deploy
 on:
   push:
-    branches: develop
+    branches: main
   pull_request:
-    branches: develop
+    branches: main
 
 jobs:
   deploy:
